Added non-working hack to try to work around Unreal building problems
It seems that Unreal does not build the plugins at all, if a binary
is already there. That makes this whole current scheme non-functional,
since currently we're compiling all Haxe code into the plugin
binary.
We'll probably need to rethink the whole scheme of where the Haxe
code gets compiled to, since it doesn't even seem that Unreal
supports live reloading of plugin code at the moment.

This current added hack will probably be deleted