Fix occasional `isValid` false positive
This happens because we used to check whether an object was valid simply
by checking the existence of its position on the object array;
However, on some cases, an object is marked as pending kill and not taken
off from the object array immediately.
This led to objects being deleted after garbage collection was done,
at which case it was too late to invalidate them on the Haxe side.
That's why there is a `IsPendingKill` check.

The fix is simple: Add a pending kill and unreachable check
so that we invalidate these cases as well.