Garbage collection and `isValid()` fixes
* Fixed the UObject array not being updated every garbage collection
since 4.19 - (due to a missing PRef<>)
* Take off bogus `objectToIndex` call on the `isValid` check - it was
referencing an old commit e3603605f0d7de540e251aef35c539d0bb2edaf2 that
mentioned that the object array could be dumped. I see no indication
that this is the case, so either it was a first try at fixing issues
when `isValid` is called on another thread (fixed by adding the
`threadSafe` option, which was renamed to `checkIfReachable` to better
illustrate what it does - since this is not exactly thread-safe as the
object might get collected in parallel)
* Add an option to debug wrappers created dynamically through
UProperties
* Fixed issues where a UProperty may get garbage collected if it's a
blueprint uproperty. Now we add to root and add a reference counter
for those properties